+++ dependency/lpp/periodic_session/assistance_data.cpp
@@ -64,12 +64,14 @@ bool AssistanceDataHandler::request_assistance_data() {
     VSCOPE_FUNCTION();
 
     messages::RequestAssistanceData message_description{};
-    message_description.cell             = mData.cell;
-    message_description.periodic_session = handle();
-    message_description.gps              = mData.gnss.gps;
-    message_description.glonass          = mData.gnss.glonass;
-    message_description.galileo          = mData.gnss.galileo;
-    message_description.bds              = mData.gnss.beidou;
+    message_description.cell               = mData.cell;
+    message_description.periodic_session   = handle();
+    message_description.gps                = mData.gnss.gps;
+    message_description.glonass            = mData.gnss.glonass;
+    message_description.galileo            = mData.gnss.galileo;
+    message_description.bds                = mData.gnss.beidou;
+    message_description.delivery_amount    = mData.config.delivery_amount;
+    message_description.rtk_antenna_height = mData.config.antenna_height;
 
     if (mData.type == RequestAssistanceData::Type::OSR) {
         message_description.rtk_observations           = 1;